requirements.txt
flask==2.3.3
pillow==10.0.0
torch==2.0.1
opencv-python==4.8.0.74
diffusers==0.21.4
transformers==4.33.1
xformers==0.0.21
numpy==1.24.3
huggingface-hub>=0.27.0
python-dotenv==1.0.0